requirements.txt
PyYAML==3.12
Logbook==1.4.3
